BATCH-671: Added a new class, JobParameter, that represents an individual parameter.  Also converted BatchStatus and ParameterType typesafe enumerations to java 5 enums.

+++ spring-batch-core/src/test/java/org/springframework/batch/core/repository/dao/AbstractJobInstanceDaoTests.java
@@ -43,8 +43,8 @@ public abstract class AbstractJobInstanceDaoTests extends AbstractTransactionalD
 		assertEquals(fooJob.getName(), retrievedInstance.getJobName());
 		assertEquals(fooParams, retrievedParams);
 		
-		assertEquals(Long.MAX_VALUE, retrievedParams.getLong("longKey").longValue());
-		assertEquals(Double.MAX_VALUE, retrievedParams.getDouble("doubleKey").doubleValue(), 0.001);
+		assertEquals(Long.MAX_VALUE, retrievedParams.getLong("longKey"));
+		assertEquals(Double.MAX_VALUE, retrievedParams.getDouble("doubleKey"), 0.001);
 		assertEquals("stringValue", retrievedParams.getString("stringKey"));
 		assertEquals(new Date(DATE), retrievedParams.getDate("dateKey"));
 	}
